#27 TaskComparater separated
TaskComparator, getPeakRSS and getCurrentRSS extracted in
new "MPI_Scheduler_Centralized_Utils" files
